done code user infomation bill
