+++ src/routes/adminRoute.ts
@@ -30,8 +30,6 @@ import Createpassword from 'src/pages/auth/forgot-password/Createpassword';
 import UpdateDeposit from 'src/pages/admin/keep-room/updateDeposit';
 import TemplateEmail from 'src/pages/admin/sendEmail';
 import CreateTemplateEmail from 'src/pages/admin/sendEmail/create-email';
-import { UserIndex } from 'src/pages/user/index';
-import ReportDetail from 'src/pages/user/reportDetail';
 import CreateAssets from 'src/pages/admin/Assets/createAssets/CreateAssets';
 import UpdateAssets from 'src/pages/admin/Assets/UpdateAssets/updateAssets';
 import UpdateService from 'src/pages/admin/service/UpdateService/updateService';
@@ -194,15 +192,6 @@ export const adminRoutes = [
     path: urlRouter.CREATE_EMAIL,
     component: CreateTemplateEmail,
   },
-  {
-    path: urlRouter.USER,
-    component: UserIndex,
-    children: [
-      { path: urlRouter.USERREPORTDETAIL, component: ReportDetail, index: true },
-
-    ],
-  },
-
   {
 
     path: urlRouter.CREATE_ASSETS,
