+++ src/routes/clientRoute.ts
@@ -1,8 +1,8 @@
 import ClientArise from "src/pages/client/arise/ClientArise";
 import ClientAsset from "src/pages/client/asset/ClientAsset";
-import ClientBill from "src/pages/client/bill/ClientBill";
+import { ClientBill } from "src/pages/client/bill/ClientBill";
 import ClientContract from "src/pages/client/contract/ClientContract";
-import ClientInfomation from "src/pages/client/information/ClientInfomation";
+import { ClientInfomation } from "src/pages/client/information/ClientInfomation";
 import ClientService from "src/pages/client/service/ClientService";
 import { urlRouter } from "src/utils/constants";
